Moved product page to user folder
